refactor: util에 있던 on과 emit events.ts로 분리
